# swagger_server/models/track.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Track(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, id: int=None, binary: bytearray=None):  # noqa: E501
        """Track - a model defined in Swagger

        :param id: The id of this Track.  # noqa: E501
        :type id: int
        :param binary: The binary of this Track.  # noqa: E501
        :type binary: bytearray
        """
        self.swagger_types = {
            'id': int,
            'binary': bytearray
        }

        self.attribute_map = {
            'id': 'id',
            'binary': 'binary'
        }
        self._id = id
        self._binary = binary

    @classmethod
    def from_dict(cls, dikt) -> 'Track':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Track of this Track.  # noqa: E501
        :rtype: Track
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> int:
        """Gets the id of this Track.


        :return: The id of this Track.
        :rtype: int
        """
        return self._id

    @id.setter
    def id(self, id: int):
        """Sets the id of this Track.


        :param id: The id of this Track.
        :type id: int
        """
        if id is None:
            raise ValueError("Invalid value for `id`, must not be `None`")  # noqa: E501

        self._id = id

    @property
    def binary(self) -> bytearray:
        """Gets the binary of this Track.


        :return: The binary of this Track.
        :rtype: bytearray
        """
        return self._binary

    @binary.setter
    def binary(self, binary: bytearray):
        """Sets the binary of this Track.


        :param binary: The binary of this Track.
        :type binary: bytearray
        """
        if binary is None:
            raise ValueError("Invalid value for `binary`, must not be `None`")  # noqa: E501

        self._binary = binary