# swagger_server/swagger/swagger.yaml
openapi: 3.0.4
info:
  title: Proveedor de Pistas (PP)
  description: Esta es la documentación para la API del Microservicio de Proveedor
    de Pistas (PP) del proyecto OverSounds.
  version: 1.0.0
externalDocs:
  description: Github access
  url: https://github.com/orgs/UniExtremadura/teams/gb02
servers:
- url: http://localhost
tags:
- name: track
  description: Access to tracks
paths:
  /track/{trackId}:
    get:
      tags:
      - track
      summary: Gets a track info by Id
      operationId: get_track
      parameters:
      - name: trackId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "400":
          description: Invalid ID supplied
        "404":
          description: Track not found
        "422":
          description: Validation exception
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - bandcamp_auth:
        - read:tracks
      x-openapi-router-controller: swagger_server.controllers.track_controller
    delete:
      tags:
      - track
      summary: Deletes a track.
      operationId: delete_track
      parameters:
      - name: trackId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: successful operation
        "204":
          description: "No Content, track deleted"
        "400":
          description: Invalid input
        "422":
          description: Validation exception
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - bandcamp_auth:
        - write:tracks
        - read:tracks
      x-openapi-router-controller: swagger_server.controllers.track_controller
    patch:
      tags:
      - track
      summary: Updates a track from the database with form data.
      operationId: update_track
      parameters:
      - name: trackId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Track"
        required: true
      responses:
        "200":
          description: Successful operation
        "400":
          description: Invalid input
        "422":
          description: Validation exception
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - bandcamp_auth:
        - write:tracks
        - read:tracks
      x-openapi-router-controller: swagger_server.controllers.track_controller
  /track/upload:
    post:
      tags:
      - track
      summary: Add a new track to the database
      operationId: add_track
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Track"
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "400":
          description: Invalid input
        "422":
          description: Validation exception
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - bandcamp_auth:
        - write:tracks
      x-openapi-router-controller: swagger_server.controllers.track_controller
components:
  schemas:
    Track:
      required:
      - binary
      - id
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        binary:
          type: string
          format: byte
          example: !!binary |-
            VTI5dVp5QlVhWFJzWlNCaWVTQkJjblJwYzNRPQ==
      example:
        binary: !!binary |-
          VTI5dVp5QlVhWFJzWlNCaWVTQkJjblJwYzNRPQ==
        id: 1
    Error:
      required:
      - code
      - message
      type: object
      properties:
        code:
          type: string
        message:
          type: string
  requestBodies:
    Track:
      description: Track object that needs to be added or updated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Track"
  securitySchemes:
    bandcamp_auth:
      type: apiKey
      name: token
      in: cookie
      x-apikeyInfoFunc: swagger_server.controllers.authorization_controller.check_bandcamp_auth